parser moved to class
